essay/requirements.txt
# Basic tools
setuptools
wheel

# Web framework
django>=4.2.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Machine learning
scikit-learn>=1.3.0
joblib>=1.3.0
tensorflow>=2.16.1
keras>=2.16.1

# OCR related
protobuf==3.20.0
paddlepaddle==2.4.2
paddleocr==2.6.0.3
shapely==2.0.1
pyclipper==1.3.0

# Language detection
langdetect>=1.0.9

# Text matching
rapidfuzz>=3.6.1
